Update resend_welcome_email() for 4.3.1+ way of doing wp_new_user_notification()
In WP 4.3, wp_new_user_notification() stopped sending passwords via email, and instead it sends a reset password link. The 2nd argument to the function is deprecated. 

There's also no longer a point to setting a new password for the user before resending the welcome email, as they will be asked to set one.